Big reorg, use EventBus
