+++ backend/package.json
@@ -9,7 +9,7 @@
     "copy-entitlements": "cp entitlements.mac.plist build/",
     "clean": "rimraf build/* dist/*",
     "build-frontend": "cd ../frontend && npm run build && cp -R build/* ../backend/build",
-    "build-backend": "electron-builder -mw",
+    "build-backend": "electron-builder -w",
     "compile": "tsc",
     "electron": "cross-env ELECTRON_START_URL=http://localhost:3000 electron .",
     "test": "NODE_ENV=test jest",
@@ -21,6 +21,7 @@
     "test-watch": "npm run test -- --watch"
   },
   "devDependencies": {
+    "@types/auto-launch": "5.0.1",
     "@types/debug": "4.1.4",
     "@types/electron": "1.6.10",
     "@types/express": "4.17.0",
@@ -45,6 +46,7 @@
   },
   "dependencies": {
     "airbrake-js": "1.6.8",
+    "auto-launch": "5.0.5",
     "axios": "0.19.0",
     "chokidar": "3.0.2",
     "cross-fetch": "3.0.4",
