+++ backend/src/ConnectdInstaller.ts
@@ -1,0 +1,7 @@
+import Installer from './Installer'
+
+export default new Installer({
+  name: 'connectd',
+  repoName: 'remoteit/connectd',
+  version: 'v4.6',
+})
