+++ backend/src/EventBus.ts
@@ -1,0 +1,5 @@
+import { EventEmitter } from 'events'
+
+class EventBus extends EventEmitter {}
+
+export default new EventBus()
