+++ backend/src/MuxerInstaller.ts
@@ -1,0 +1,7 @@
+import Installer from './Installer'
+
+export default new Installer({
+  name: 'muxer',
+  repoName: 'remoteit/multiport',
+  version: 'v0.3-alpha',
+})
