+++ backend/src/remote.it.ts
@@ -1,5 +1,5 @@
 import axios from 'axios'
 import setup from 'remote.it'
-import { API_URL, DEVELOPER_KEY } from '../constants'
+import { API_URL, DEVELOPER_KEY } from './constants'
 
 export const r3 = setup({ apiURL: API_URL, developerKey: DEVELOPER_KEY })
