+++ backend/src/types.d.ts
@@ -3,9 +3,30 @@ import { IService } from 'remote.it'
 import httpProxy from 'http-proxy'
 
 declare global {
-  export interface User {
-    authHash: string
+  interface UserCredentials {
     username: string
-    language?: string // 'en' | 'jp'
+    authHash: string
+  }
+
+  interface ConnectionData {
+    id: string
+    port: number
+    name?: string
+    pid?: number
+  }
+
+  interface ConnectionArgs {
+    authHash: string
+    host?: string
+    id: string
+    name?: string
+    port?: number
+    username: string
+  }
+
+  interface SavedConnection {
+    id: string
+    port: number
+    name?: string
   }
 }
