--- backend/src/utils/toPercent.ts
@@ -1,3 +1,0 @@
-export function toPercent(percent: number) {
-  return Math.floor(percent * 100)
-}
